+++ test/test_json_reader.c
@@ -22,6 +22,19 @@ typedef enum {
 } flag_type;
 
 static void test_read_file(const char *path, flag_type type, expect_type expect) {
+    
+#if YYJSON_DISABLE_UTF8_VALIDATION
+    {
+        u8 *dat;
+        usize dat_len;
+        if (yy_file_read(path, &dat, &dat_len)) {
+            bool is_utf8 = yyb_str_is_utf8((const char *)dat, dat_len);
+            free(dat);
+            if (!is_utf8) return;
+        }
+    }
+#endif
+    
     yyjson_read_flag flag = YYJSON_READ_NOFLAG;
     if (type & FLAG_COMMA) flag |= YYJSON_READ_ALLOW_TRAILING_COMMAS;
     if (type & FLAG_COMMENT) flag |= YYJSON_READ_ALLOW_COMMENTS;